requirements.txt
PyYAML==3.11
buildbot==0.8.8
buildbot-slave==0.8.8
coverage==3.7
mock==1.0.1
webrunner==0.1
pep8==1.5.6
pyflakes==0.8.1
